Main/Appliance: Allow users to specify a different storage controller and/or controller port for hard disks when importing a VM. bugref:5027
Follow-up build-fix due to a missing cast when assigning an unsigned
integer to a signed integer (flagged by [-Werror,-Wsign-conversion]).


git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@93481 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Main/src-server/ApplianceImplImport.cpp
@@ -5747,7 +5747,7 @@ l_skipped:
 
                             if (uOrigControllerPortValue != uNewControllerPortValue)
                             {
-                                AD.lPort = uNewControllerPortValue;
+                                AD.lPort = (int32_t)uNewControllerPortValue;
                             }
                             if (uOrigControllerIndex != uTargetControllerIndex)
                             {
